refactor

+++ server/controllers/category.js
@@ -4,17 +4,22 @@ import Task from '../models/task.js';
 
 export async function createCategory(req, res) {
   const { name } = req.body;
+  // make sure request is valid
+  if (!name) {
+    return res.status(400).json({
+      error: 'bad request',
+      message: 'category name missing in request',
+    });
+  }
   try {
     const category = await Category.findOne({ name, userId: req.user._id });
-
     // if the user already have a category with this name
     if (category) {
-      console.log('Heere bay');
-      return res
-        .status(422)
-        .json({ errors: ['A category with that name already exiist'] });
+      return res.status(409).json({
+        error: 'Duplicate content',
+        message: 'Category with that name already exist',
+      });
     }
-
     // create new category and add it to the category list of the user
     const newCategory = new Category({ name, userId: req.user.userId });
     const user = await Profile.findById(req.user.userId);
@@ -23,28 +28,49 @@ export async function createCategory(req, res) {
     await user.save();
     return res.status(200).json(result);
   } catch (error) {
-    return res.status(500).json(error);
+    return res.status(500).json({
+      error: 'Internal Server Error',
+      message: 'Something went wrong in the server',
+    });
   }
 }
 
+// get all category for a user
 export async function getCategories(req, res) {
   try {
     const categories = await Category.find({ userId: req.user.userId });
     return res.status(200).json(categories);
   } catch (error) {
-    console.log(error);
-    return res.status(500).json({ errors: error });
+    return res.status(500).json({
+      error: 'Internal Server Error',
+      message: 'Something went wrong in the server',
+    });
   }
 }
 
+// delete a user's category
 export async function deleteCategory(req, res) {
   const { id } = req.params;
   try {
     // delete the catetory and the tasks associated with it
-    await Category.findByIdAndDelete(id);
+    const category = await Category.findOneAndDelete({
+      _id: id,
+      userId: req.user.userId,
+    });
+
+    // verify that this category existed
+    if (!category) {
+      return res.status(404).json({
+        error: 'not found error',
+        message: 'This category was not found in our server',
+      });
+    }
     await Task.deleteMany({ category: id });
-    res.status(200).json({ message: 'Category was deleted' });
+    return res.status(200).json({ message: 'Category was deleted' });
   } catch (error) {
-    res.status(500).json({ errors: ['Something went wrong'] });
+    return res.status(500).json({
+      error: 'Internal Server Error',
+      message: 'Something went wrong on the server',
+    });
   }
 }
